pyproject.toml
[project]
name = "soulmate-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
